+++ src/main/java/com/flyn/eeg_receiver/listener/DeviceListener.java
@@ -1,0 +1,4 @@
+package com.flyn.eeg_receiver.listener;
+
+public interface DeviceListener {
+}