+++ src/main/kotlin/com/flyn/eeg_receiver/event/AttentionEvent.kt
@@ -1,4 +1,4 @@
-package event
+package com.flyn.eeg_receiver.event
 
 class AttentionEvent(val value: Int): DeviceEvent {
 }