+++ src/main/kotlin/com/flyn/eeg_receiver/event/DeviceEvent.kt
@@ -1,0 +1,4 @@
+package com.flyn.eeg_receiver.event
+
+interface DeviceEvent {
+}