+++ CHANGELOG.md
@@ -5,8 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.0] - 2021-10-08
 ### Added
 - EEG raw data viewer
+### Changed
+- move all file to com.flyn.eeg_receiver folder
+- MindSetReceiver class renamed EEGReceiver class
 
 ## [0.2.2] - 2021-10-01
 ### Fixed